+++ src/main/java/insane96mcp/survivalreimagined/module/hungerhealth/NoHunger.java
@@ -16,7 +16,6 @@ import insane96mcp.survivalreimagined.module.hungerhealth.fooddrinks.FoodDrinks;
 import insane96mcp.survivalreimagined.module.movement.stamina.Stamina;
 import insane96mcp.survivalreimagined.network.NetworkHandler;
 import insane96mcp.survivalreimagined.network.message.FoodRegenSyncMessage;
-import insane96mcp.survivalreimagined.setup.SRMobEffects;
 import insane96mcp.survivalreimagined.utils.ClientUtils;
 import insane96mcp.survivalreimagined.utils.LogHelper;
 import insane96mcp.survivalreimagined.utils.Utils;
@@ -191,8 +190,8 @@ public class NoHunger extends Feature {
         secs = (int) ((passiveRegenerationTime.max - passiveRegenerationTime.min) * healthPerc + passiveRegenerationTime.min);
         if (player.level().getDifficulty().equals(Difficulty.HARD))
             secs *= 1.5d;
-        if (player.hasEffect(SRMobEffects.WELL_FED.get())) {
-            MobEffectInstance wellFed = player.getEffect(SRMobEffects.WELL_FED.get());
+        if (player.hasEffect(HealthRegen.WELL_FED.get())) {
+            MobEffectInstance wellFed = player.getEffect(HealthRegen.WELL_FED.get());
             //noinspection ConstantConditions
             secs *= 1 - (((wellFed.getAmplifier() + 1) * 0.4d));
         }