+++ src/main/java/insane96mcp/survivalreimagined/setup/Strings.java
@@ -4,13 +4,12 @@ import insane96mcp.survivalreimagined.SurvivalReimagined;
 
 import java.util.UUID;
 
+//TODO Remove
 public class Strings {
 	public static class Tags {
 		public static final String SPAWNED_MOBS = SurvivalReimagined.RESOURCE_PREFIX + "spawned_mobs";
 
 		public static final String SPAWNER_DATA = SurvivalReimagined.RESOURCE_PREFIX + "spawner_data";
-		public static final String DAMAGE_HISTORY = SurvivalReimagined.RESOURCE_PREFIX + "damage_history";
-		public static final String EAT_HISTORY = SurvivalReimagined.RESOURCE_PREFIX + "eat_history";
         public static final String PASSIVE_REGEN_TICK = SurvivalReimagined.RESOURCE_PREFIX + "passive_regen_ticks";
 		public static final String SPAWNER_DISABLED = SurvivalReimagined.RESOURCE_PREFIX + "spawner_disabled";
     }
@@ -18,9 +17,6 @@ public class Strings {
 	public static class AttributeModifiers {
 		public static final UUID ARMOR_SLOWDOWN_UUID = UUID.fromString("8588420e-ce50-4e4e-a3e4-974dfc8a98ec");
 		public static final String ARMOR_SLOWDOWN = SurvivalReimagined.RESOURCE_PREFIX + "armor_slowdown";
-		public static final UUID WEATHER_SLOWDOWN_UUID = UUID.fromString("81cd06ed-b645-4c96-91fe-e29e940503c3");
-		public static final String WEATHER_SLOWDOWN = SurvivalReimagined.RESOURCE_PREFIX + "weather_slowdown";
-		public static final UUID GENERIC_ITEM_MODIFIER_UUID = UUID.fromString("8ba29557-f4dd-449d-a1b8-396ed980d042");
 		public static final String GENERIC_ITEM_MODIFIER = SurvivalReimagined.RESOURCE_PREFIX + "item_modifier";
 	}